refac: Move exception-related code to new exception subproject
