+++ apps/payment/src/main/java/com/github/thorlauridsen/controller/IPaymentController.java
@@ -1,7 +1,7 @@
 package com.github.thorlauridsen.controller;
 
-import com.github.thorlauridsen.dto.ErrorDto;
 import com.github.thorlauridsen.dto.PaymentDto;
+import com.github.thorlauridsen.exception.ErrorDto;
 import com.github.thorlauridsen.exception.PaymentNotFoundException;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
