+++ modules/exception/src/main/java/com/github/thorlauridsen/exception/ControllerAdvisor.java
@@ -1,6 +1,6 @@
 package com.github.thorlauridsen.exception;
 
-import com.github.thorlauridsen.dto.ErrorDto;
+import java.time.OffsetDateTime;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
@@ -8,8 +8,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-
-import java.time.OffsetDateTime;
 
 /**
  * Controller advisor for handling exceptions.
