+++ modules/exception/src/main/java/com/github/thorlauridsen/exception/DomainException.java
